v0.0 build
Create file functions py
