+++ aa_resourcecog/player.py
@@ -7,6 +7,7 @@ from itertools import chain
 
 #from .aa_resourcecog import clash_embed, alliance_file_handler, data_file_handler
 from .constants import emotes_townhall, emotes_heroes, hero_availability, troop_availability, spell_availability
+from .file_functions import get_current_alliance, season_file_handler, alliance_file_handler, data_file_handler
 
 from .notes import aNote
 from .clan import aClan
